pyproject.toml
[project]
name = "browsing-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.6",
    "autogen-agentchat>=0.6.4",
    "autogen-ext[azure,mcp,openai]>=0.6.4",
    "browser-use>=0.5.4",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
]
